Each adapter can now reference other adapter types.
This addition makes changes so that adapters can take other adapters
as a parameter. The reason for this change is because it is needed
for logic adapters. Logic adapters should be able to access the
storage adapter to retrieve information. It is more efficient to
allow logic adapters direct access to the database than it is to
pass in a list of known statements as is currently done. With this
in mind, I have given all adapters access to each other to keep
the interfaces unified and easy to work with.